Added validation

+++ src/main/java/ru/sber/appointment/controller/AdminController.java
@@ -2,6 +2,7 @@ package ru.sber.appointment.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import ru.sber.appointment.model.Doctor;
 import ru.sber.appointment.model.User;
@@ -9,6 +10,10 @@ import ru.sber.appointment.service.AuthServiceImpl;
 import ru.sber.appointment.service.DoctorServiceImpl;
 import ru.sber.appointment.service.ScheduleServiceImpl;
 import ru.sber.appointment.service.UserServiceImpl;
+import ru.sber.appointment.utils.Validator;
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 
 @RestController
 @RequestMapping("api/v1/admin/{username}")
@@ -16,7 +21,8 @@ public class AdminController {
 
     @Autowired
     UserServiceImpl userService;
-
+    @Autowired
+    Validator validator;
     @Autowired
     AuthServiceImpl authService;
 
@@ -26,27 +32,43 @@ public class AdminController {
     DoctorServiceImpl doctorService;
 
     @DeleteMapping("delete/user")
-    public ResponseEntity<?> deleteUser(@PathVariable String username, @RequestBody User user){
-        userService.deleteUser(user);
-        return ResponseEntity.ok("OK");
+    public ResponseEntity<?> deleteUser(@NotNull @PathVariable String username, @RequestBody User user, BindingResult bindingResult){
+        if (bindingResult.hasErrors()) {
+            return validator.getErrorList(bindingResult);
+        } else {
+            userService.deleteUser(user);
+            return ResponseEntity.ok("OK");
+        }
     }
 
     @DeleteMapping("delete/doctor")
-    public ResponseEntity<?> deleteDoctor(@PathVariable String username, @RequestBody Doctor doctor){
-        doctorService.deleteDoctor(doctor);
-        return ResponseEntity.ok("OK");
+    public ResponseEntity<?> deleteDoctor(@NotNull @PathVariable String username, @Valid @RequestBody Doctor doctor, BindingResult bindingResult){
+        if (bindingResult.hasErrors()) {
+            return validator.getErrorList(bindingResult);
+        } else {
+            doctorService.deleteDoctor(doctor);
+            return ResponseEntity.ok("OK");
+        }
     }
 
     @PostMapping("post/doctor")
-    public ResponseEntity<?> saveDoctor(@PathVariable String username, @RequestBody Doctor doctor){
-        doctorService.saveDoctor(doctor);
-        return ResponseEntity.ok("OK");
+    public ResponseEntity<?> saveDoctor(@NotNull @PathVariable String username, @Valid @RequestBody Doctor doctor , BindingResult bindingResult){
+        if (bindingResult.hasErrors()) {
+            return validator.getErrorList(bindingResult);
+        } else {
+            doctorService.saveDoctor(doctor);
+            return ResponseEntity.ok("OK");
+        }
     }
 
     @PostMapping("post/schedule/bulk")
-    public ResponseEntity<?> saveScheduleForWeek(@PathVariable String username, @RequestBody Doctor doctor){
-        scheduleService.saveScheduleForWeek(doctor);
-        return ResponseEntity.ok("OK");
+    public ResponseEntity<?> saveScheduleForWeek(@NotNull @PathVariable String username, @Valid @RequestBody Doctor doctor , BindingResult bindingResult){
+        if (bindingResult.hasErrors()) {
+            return validator.getErrorList(bindingResult);
+        } else {
+            scheduleService.saveScheduleForWeek(doctor);
+            return ResponseEntity.ok("OK");
+        }
     }
 
     @GetMapping("/get")
